refactor: AI型定義を論理的なファイルに分割し組織化を改善
- types.tsの大きなファイルを論理的なグループに分割
- AISettings.ts: AI設定関連の型定義
- GameState.ts: ゲーム状態関連の型定義（AIFieldState, AIPuyoPairState, AIGameState）
- MoveTypes.ts: 移動・評価関連の型定義（AIMove, PossibleMove, EvaluationResult, MoveEvaluation）
- PerformanceTypes.ts: パフォーマンス分析関連の型定義（GameSession, PerformanceData, PerformanceReport）
- index.ts: 統合エクスポートファイルで全型定義をre-export
- types.ts: 後方互換性維持のためのre-exportファイルに変更
- 既存のimport文を壊すことなく段階的移行を可能にする設計
- コードの可読性と保守性が向上
